Fix focus management for cacheable popups on RN Windows (#606)
* Refactor /web/PopupContainer.tsx

- Rename web/PopupContainer.tsx to web/PopupContainerView.tsx for consistency
with native-common.
- Pull the popup component tracking functionality in web/PopupContainerView.tsx
into a new base class common/PopupContainerViewBase.tsx.

* Fix focus management for cacheable popups on RN Windows

Just like web, Windows views rendered into a cacheable popup must be aware of the popup
hiding and re-showing in order to properly un-restrict and re-restrict focus to the
right components. The implementation is identical to that used on web, i.e. FocusManager
is released and re-initialized on hide and re-show, respectively, in addition to doing
it on unmount and mount.

* Feedback: Type safety and naming conventions

(cherry picked from commit 8e4602ad3249bbfd3b4453c1799208a771a2318e)