Use zero-sized PopupContainerView for RN hidden popups (#607)
Passing 0 for size to renderPopup() is not enough to make the popup reliably
render as hidden. The bug reproduces as "ghost" components showing on the screen
when a regular popup is shown and a cacheable popup is rendered as hidden.

(cherry picked from commit cccbb0b33a2c993afff7da5bfb664ce2b1efbac8)