Always take View ref in PopupContainerView (#616)
The old code was likely an attempt at optimizing out the measurement/position logic for
hidden popups. It's not only unnecessary (the code is under if !this.props.hidden already)
but also wrong because if a hidden popup happens to get unmounted, _mountedComponent will
not be updated and we'll end up calling setState on an unmounted component.

(cherry picked from commit a3f07ee6a23ffc31947956c199a06a62d41bfbcf)